% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\name{.onLoad}
\alias{.onLoad}
\title{Global Variable Definitions}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Path to the package library.}

\item{pkgname}{Package name.}
}
\description{
This file defines the necessary global variables used for tidy
evaluation (\code{.data}) to prevent "no visible binding" NOTES during \code{R CMD check}.
}
\keyword{internal}
