% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_species.R
\name{get_species_code}
\alias{get_species_code}
\title{Get Species Code}
\usage{
get_species_code(commonName)
}
\arguments{
\item{commonName}{Case-sensitive common name.}
}
\value{
Integer species code.
}
\description{
Looks up the numeric species code for a given common name.
}
