% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coverage.R
\name{map_coverage}
\alias{map_coverage}
\title{Generate Atlas Effort (Coverage) Map}
\usage{
map_coverage(country, scale = "large", save_dir = getwd())
}
\arguments{
\item{country}{A character string specifying the country name (e.g., "South Africa").}

\item{scale}{A character string indicating the scale of the natural earth boundaries.
?? (e.g., "large", "medium"). Defaults to "large".}

\item{save_dir}{A character string specifying the directory to save the output PNG file.
?? Defaults to the current working directory.}
}
\value{
The file path to the generated map image (invisibly).
}
\description{
Downloads, processes, and plots pentad-level coverage effort
?? (Total Hours) for a specified country from the African Bird Atlas Project (ABAP) API.
}
