% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_codes.R
\name{show_codes}
\alias{show_codes}
\title{Show Species Codes}
\usage{
show_codes(species_codes)
}
\arguments{
\item{species_codes}{A numeric vector of species codes.}
}
\value{
Prints the number of codes and lists them.
}
\description{
A utility to review a list of species codes before batch processing.
Checks if the list exceeds the batch limit of 10.
}
