% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Species Codes and Names}
\format{
A data frame with columns:
\describe{
\item{species_code}{The unique numeric identifier for the species}
\item{commonName}{The English common name of the species}
\item{scientificName}{The scientific name of the species}
}
}
\usage{
species_data
}
\description{
A dataset containing common names and numeric codes for bird species.
}
\keyword{datasets}
