% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species_code}
\alias{get_species_code}
\title{Get Species Code}
\usage{
get_species_code(species_name)
}
\arguments{
\item{species_name}{A character string representing either the case-sensitive
common name or the scientific name.}
}
\value{
Integer species code. Returns \code{NA_integer_} if not found.
}
\description{
Looks up the numeric species code for a given common name or scientific name.
}
